requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.11
mysqlclient==2.0.3
protobuf==3.17.3
six==1.16.0
Werkzeug==2.0.1
